DumpFile can now handle BigInteger offsets

+++ investigations/migrations/0001_initial.py
@@ -50,7 +50,7 @@ class Migration(migrations.Migration):
             name='FileDump',
             fields=[
                 ('file_dump_id', models.AutoField(primary_key=True, serialize=False)),
-                ('offset', models.IntegerField()),
+                ('offset', models.BigIntegerField()),
                 ('filename', models.CharField(max_length=255)),
                 ('case_id', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, to='investigations.uploadinvestigation')),
                 ('is_malicious', models.BooleanField(default=False)),
